+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Enhancements
 
 ### Bug Fixes
+- Unblock Search Relevance requests on legacy system index mapping conflicts ([#599](https://github.com/opensearch-project/search-relevance/pull/599))
 
 ### Infrastructure